refactor: create pages folder and replace all route components
