--- src/pages/Main.tsx
@@ -1,7 +1,0 @@
-import "./main.module.scss";
-
-const Main = () => {
-  return <h1>Hello world</h1>;
-};
-
-export default Main;
