--- src/pages/main.module.scss
@@ -1,3 +1,0 @@
-h1 {
-  font-size: 50px;
-}
