+++ src/pages/todoBoard/TodoBoard.tsx
@@ -1,10 +1,12 @@
-import Header from "../Header/Header";
+import Header from "../../components/Header/Header";
+import TodoContainer from "../../containers/TodoContainer/TodoContainer";
 import classes from "./todoBoard.module.scss";
 
 const TodoBoard = () => {
   return (
     <div className={classes.todoContainer}>
       <Header />
+      <TodoContainer />
     </div>
   );
 };
